Likes List: load more Likes on scroll

+++ WordPress/Classes/Services/CommentService+Likes.swift
@@ -3,17 +3,20 @@ extension CommentService {
     /**
      Fetches a list of users from remote that liked the comment with the given IDs.
      
-     @param commentID   The ID of the comment to fetch likes for
-     @param siteID      The ID of the site that contains the post
-     @param count       Number of records to retrieve. Optional. Defaults to the endpoint max of 90.
-     @param before      Filter results to likes before this date/time. Optional.
-     @param success     A success block
-     @param failure     A failure block
+     @param commentID       The ID of the comment to fetch likes for
+     @param siteID          The ID of the site that contains the post
+     @param count           Number of records to retrieve. Optional. Defaults to the endpoint max of 90.
+     @param before          Filter results to likes before this date/time. Optional.
+     @param purgeExisting   Indicates if existing Likes for the given post and site should be purged before
+                            new ones are created. Defaults to false.
+     @param success         A success block
+     @param failure         A failure block
      */
     func getLikesFor(commentID: NSNumber,
                      siteID: NSNumber,
                      count: Int = 90,
                      before: String? = nil,
+                     purgeExisting: Bool = false,
                      success: @escaping (([LikeUser], Int) -> Void),
                      failure: @escaping ((Error?) -> Void)) {
 
@@ -23,13 +26,15 @@ extension CommentService {
             return
         }
 
-
         remote.getLikesForCommentID(commentID,
                                     count: NSNumber(value: count),
                                     before: before,
                                     success: { remoteLikeUsers, totalLikes in
-                                        self.createNewUsers(from: remoteLikeUsers, commentID: commentID, siteID: siteID) {
-                                            let users = self.likeUsersFor(commentID: commentID, siteID: siteID)
+                                        self.createNewUsers(from: remoteLikeUsers,
+                                                            commentID: commentID,
+                                                            siteID: siteID,
+                                                            purgeExisting: purgeExisting) {
+                                            let users = self.likeUsersFor(commentID: commentID, siteID: siteID, before: before)
                                             success(users, totalLikes.intValue)
                                             LikeUserHelper.purgeStaleLikes()
                                         }
@@ -42,12 +47,21 @@ extension CommentService {
     /**
      Fetches a list of users from Core Data that liked the comment with the given IDs.
      
-     @param commentID  The ID of the comment to fetch likes for
-     @param siteID     The ID of the site that contains the post
+     @param commentID   The ID of the comment to fetch likes for.
+     @param siteID      The ID of the site that contains the post.
+     @param before      Filter results to likes before this date/time. Optional.
      */
-    func likeUsersFor(commentID: NSNumber, siteID: NSNumber) -> [LikeUser] {
+    func likeUsersFor(commentID: NSNumber, siteID: NSNumber, before: String? = nil) -> [LikeUser] {
         let request = LikeUser.fetchRequest() as NSFetchRequest<LikeUser>
-        request.predicate = NSPredicate(format: "likedSiteID = %@ AND likedCommentID = %@", siteID, commentID)
+
+        request.predicate = {
+            if let beforeDate = DateUtils.date(fromISOString: before) {
+                return NSPredicate(format: "likedSiteID = %@ AND likedCommentID = %@ AND dateLiked < %@", siteID, commentID, beforeDate as CVarArg)
+            }
+
+            return NSPredicate(format: "likedSiteID = %@ AND likedCommentID = %@", siteID, commentID)
+        }()
+
         request.sortDescriptors = [NSSortDescriptor(key: "dateLiked", ascending: false)]
 
         if let users = try? managedObjectContext.fetch(request) {
@@ -64,6 +78,7 @@ private extension CommentService {
     func createNewUsers(from remoteLikeUsers: [RemoteLikeUser]?,
                         commentID: NSNumber,
                         siteID: NSNumber,
+                        purgeExisting: Bool,
                         onComplete: @escaping (() -> Void)) {
 
         guard let remoteLikeUsers = remoteLikeUsers,
@@ -76,7 +91,9 @@ private extension CommentService {
 
         derivedContext.perform {
 
-            self.deleteExistingUsersFor(commentID: commentID, siteID: siteID, from: derivedContext)
+            if purgeExisting {
+                self.deleteExistingUsersFor(commentID: commentID, siteID: siteID, from: derivedContext)
+            }
 
             remoteLikeUsers.forEach {
                 LikeUserHelper.createUserFrom(remoteUser: $0, context: derivedContext)
